+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * [FEATURE] Added a CTA Section
 * [FEATURE] Added a Features Section
+* [FEATURE] Added a Footer Section
 * [FEATURE] Added a Hero Section
 * [FEATURE] Added a Nav Bar
 * [FEATURE] Added a Pricing Section