+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * [FEATURE] Added a Hero Section
 * [FEATURE] Added a Nav Bar
 * [FEATURE] Added a Pricing Section
+* [FEATURE] Added a Privacy Policy Page
 * [FEATURE] Added a Product Image Section
 * [UI] Added a Tailwind Color Theme
 * [UI] Added a Download images for Google, Apple and Shopify